+++ src/Analyzers/CSharp/DiagnosticRules.Generated.cs
@@ -2639,5 +2639,17 @@ namespace Roslynator.CSharp
             helpLinkUri:        DiagnosticIdentifiers.UseRawStringLiteral, 
             customTags:         Array.Empty<string>());
 
+        /// <summary>RCS1267</summary>
+        public static readonly DiagnosticDescriptor UseStringInterpolationInsteadOfStringConcat = DiagnosticDescriptorFactory.Create(
+            id:                 DiagnosticIdentifiers.UseStringInterpolationInsteadOfStringConcat, 
+            title:              "Use string interpolation instead of 'string.Concat'", 
+            messageFormat:      "Use string interpolation instead of 'string.Concat'", 
+            category:           DiagnosticCategories.Roslynator, 
+            defaultSeverity:    DiagnosticSeverity.Info, 
+            isEnabledByDefault: true, 
+            description:        null, 
+            helpLinkUri:        DiagnosticIdentifiers.UseStringInterpolationInsteadOfStringConcat, 
+            customTags:         Array.Empty<string>());
+
     }
 }